python_engine/crawler_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
pyyaml==6.0.1
apscheduler==3.10.4
httpx==0.25.2
motor>=3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
